Fixed bug in AdvertentieDAO Created class Reactie Removed save method from DAOInterface

+++ src/main/java/nl/corne/marktplaats/model/advertentie/Advertentie.java
@@ -25,7 +25,7 @@ import java.math.BigDecimal;
 public class Advertentie {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int advID;
+    private long advID;
     @ManyToOne(cascade = CascadeType.PERSIST)
     private Gebruiker mygebruiker;
     @Enumerated
